Projects: Move FilteredDirView to its widgets
