+++ spyder/plugins/explorer/widgets/explorer.py
@@ -1687,90 +1687,6 @@ class ProxyModel(QSortFilterProxyModel):
         return self.sourceModel().type(self.mapToSource(index))
 
 
-class FilteredDirView(DirView):
-    """Filtered file/directory tree view."""
-    def __init__(self, parent=None):
-        """Initialize the filtered dir view."""
-        super().__init__(parent)
-        self.proxymodel = None
-        self.setup_proxy_model()
-        self.root_path = None
-
-    # ---- Model
-    def setup_proxy_model(self):
-        """Setup proxy model."""
-        self.proxymodel = ProxyModel(self)
-        self.proxymodel.setSourceModel(self.fsmodel)
-
-    def install_model(self):
-        """Install proxy model."""
-        if self.root_path is not None:
-            self.setModel(self.proxymodel)
-
-    def set_root_path(self, root_path):
-        """
-        Set root path.
-
-        Parameters
-        ----------
-        root_path: str
-            New path directory.
-        """
-        self.root_path = root_path
-        self.install_model()
-        index = self.fsmodel.setRootPath(root_path)
-        self.proxymodel.setup_filter(self.root_path, [])
-        self.setRootIndex(self.proxymodel.mapFromSource(index))
-
-    def get_index(self, filename):
-        """
-        Return index associated with filename.
-
-        Parameters
-        ----------
-        filename: str
-            String with the filename.
-        """
-        index = self.fsmodel.index(filename)
-        if index.isValid() and index.model() is self.fsmodel:
-            return self.proxymodel.mapFromSource(index)
-
-    def set_folder_names(self, folder_names):
-        """
-        Set folder names
-
-        Parameters
-        ----------
-        folder_names: list
-            List with the folder names.
-        """
-        assert self.root_path is not None
-        path_list = [osp.join(self.root_path, dirname)
-                     for dirname in folder_names]
-        self.proxymodel.setup_filter(self.root_path, path_list)
-
-    def get_filename(self, index):
-        """
-        Return filename from index
-
-        Parameters
-        ----------
-        index: int
-            Index of the list of filenames
-        """
-        if index:
-            path = self.fsmodel.filePath(self.proxymodel.mapToSource(index))
-            return osp.normpath(str(path))
-
-    def setup_project_view(self):
-        """Setup view for projects."""
-        for i in [1, 2, 3]:
-            self.hideColumn(i)
-        self.setHeaderHidden(True)
-        # Disable the view of .spyproject.
-        self.filter_directories()
-
-
 class ExplorerTreeWidget(DirView):
     """
     File/directory explorer tree widget.
